filtering/requirements.txt
fire
numpy
peft
pyyaml
torch
tqdm
transformers
loguru
matplotlib
